+++ src/gen_decompile.ml
@@ -181,8 +181,10 @@ let to_michelson (input, env : T.obj_micheline * env) : T.michelson * env =
       | Oprim ({prim = "SOURCE"; _})                         -> T.SOURCE
       | Oprim ({prim = "TRANSFER_TOKENS"; _})                -> T.TRANSFER_TOKENS
       (* Operations on data structures *)
-      | Oprim ({prim = "CAR"; _})                            -> T.CAR
-      | Oprim ({prim = "CDR"; _})                            -> T.CDR
+      | Oprim ({prim = "CAR"; args = n::_})                  -> T.CAR (to_int n)
+      | Oprim ({prim = "CAR"; _})                            -> T.CAR 0
+      | Oprim ({prim = "CDR"; args = n::_})                  -> T.CDR (to_int n)
+      | Oprim ({prim = "CDR"; _})                            -> T.CDR 1
       | Oprim ({prim = "CONCAT"; _})                         -> T.CONCAT
       | Oprim ({prim = "CONS"; _})                           -> T.CONS
       | Oprim ({prim = "EMPTY_BIG_MAP" ; args = k::v::_})    -> T.EMPTY_BIG_MAP (to_type k, to_type v)
@@ -473,12 +475,12 @@ let to_dir (michelson, env : T.michelson * env) =
           | T.Dbop (Bpair, a1, b1), _ -> begin
               let f op =
                 match b, op with
-                | T.Dbop (Bpair, x, _), T.Ucar -> x
-                | T.Dbop (Bpair, _, y), T.Ucdr -> y
+                | T.Dbop (Bpair, x, _), T.Ucar _ -> x
+                | T.Dbop (Bpair, _, y), T.Ucdr _ -> y
                 | _ -> T.Duop (op, b)
               in
-              let car = f Ucar in
-              let cdr = f Ucdr in
+              let car = f (Ucar 0) in
+              let cdr = f (Ucdr 1) in
               aux (aux accu (a1, car)) (b1, cdr)
             end
           | _ -> (a, b)::accu
@@ -531,8 +533,8 @@ let to_dir (michelson, env : T.michelson * env) =
                       List.map (fun instr -> begin
                             let rec fe (expr : T.dexpr) =
                               match expr with
-                              | T.Duop (T.Ucar, T.Dbop (Bpair, x, _)) -> fe x
-                              | T.Duop (T.Ucdr, T.Dbop (Bpair, _, y)) -> fe y
+                              | T.Duop (T.Ucar 0, T.Dbop (Bpair, x, _)) -> fe x
+                              | T.Duop (T.Ucdr 1, T.Dbop (Bpair, _, y)) -> fe y
                               | _ -> T.map_dexpr fe id id expr
                             in
                             let rec f (x : T.dinstruction) : T.dinstruction = T.map_dexpr_dinstr f fe id x in
@@ -933,8 +935,8 @@ let to_dir (michelson, env : T.michelson * env) =
 
     (* Operations on data structures *)
 
-    | CAR::it                  -> interp_uop env Ucar it stack
-    | CDR::it                  -> interp_uop env Ucdr it stack
+    | CAR n::it               -> interp_uop env (Ucar n) it stack
+    | CDR n::it               -> interp_uop env (Ucdr n) it stack
     | CONCAT::it               -> interp_bop env Bconcat it stack
     | CONS::it                 -> interp_bop env Bcons it stack
     | EMPTY_BIG_MAP (k, v)::it -> interp_zop env (Zemptybigmap (k, v)) it stack
@@ -1426,8 +1428,8 @@ let to_model (ir, env : T.ir * env) : M.model * env =
       end
     | Iunop (op, e) -> begin
         match op with
-        | Ucar               -> M.mk_mterm (Mtupleaccess (f e, Big_int.zero_big_int))  (M.tunit)
-        | Ucdr               -> M.mk_mterm (Mtupleaccess (f e, Big_int.unit_big_int))  (M.tunit)
+        | Ucar _n            -> M.mk_mterm (Mtupleaccess (f e, Big_int.zero_big_int)) (M.tunit)
+        | Ucdr _n            -> M.mk_mterm (Mtupleaccess (f e, Big_int.unit_big_int)) (M.tunit)
         | Uleft  t           -> let ee = f e in let t = for_type t in M.mk_mterm (Mleft  (t, f e)) (M.tor ee.type_ t)
         | Uright t           -> let ee = f e in let t = for_type t in M.mk_mterm (Mright (t, f e)) (M.tor t ee.type_)
         | Uneg               -> M.mk_mterm (Muminus (f e)) M.tint